// oficina-mecanica/src/types/Professionals.tsx
export type Professionals = {
    id: number;
    name: string;
    birth: Date;
    firstDateContract: Date;
    endDateContract: Date;
    fone: number;
    cep: number;
    nickname: string;
    email: string;
    cpf: number;
    address: string;
    number_address: number;
    state: string;
    obs: string;
}

  
export interface Employee {
    id: number;
    name: string;
    birth: string;
    firstDateContract: string;
    endDateContract: string;
    fone: number;
    cep: number;
    nickname: string;
    email: string;
    cpf: number;
    address: string;
    numberAddress: number;
    state: string;
    obs: string;
  }